+++ R/ngram-tokenizers.R
@@ -78,11 +78,11 @@ tokenize_ngrams <- function(x, lowercase = TRUE, n = 3L, n_min = n,
 #' @export
 #' @rdname ngram-tokenizers
 tokenize_skip_ngrams <- function(x, lowercase = TRUE, n = 3, k = 1,
-                                 simplify = FALSE) {
+                                 stopwords = character(), simplify = FALSE) {
   check_input(x)
   named <- names(x)
   words <- tokenize_words(x, lowercase = lowercase)
-  out <- lapply(words, skip_ngrams, n = n, k = k)
+  out <- skip_ngrams_vectorised(words, stopwords, n, k)
   if (!is.null(named)) names(out) <- named
   simplify_list(out, simplify)
 }