+++ src/RcppExports.cpp
@@ -20,16 +20,17 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// skip_ngrams
-CharacterVector skip_ngrams(CharacterVector words, int n, int k);
-RcppExport SEXP tokenizers_skip_ngrams(SEXP wordsSEXP, SEXP nSEXP, SEXP kSEXP) {
+// skip_ngrams_vectorised
+ListOf<CharacterVector> skip_ngrams_vectorised(ListOf<CharacterVector> words, CharacterVector stopwords, int n, int k);
+RcppExport SEXP tokenizers_skip_ngrams_vectorised(SEXP wordsSEXP, SEXP stopwordsSEXP, SEXP nSEXP, SEXP kSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< CharacterVector >::type words(wordsSEXP);
+    Rcpp::traits::input_parameter< ListOf<CharacterVector> >::type words(wordsSEXP);
+    Rcpp::traits::input_parameter< CharacterVector >::type stopwords(stopwordsSEXP);
     Rcpp::traits::input_parameter< int >::type n(nSEXP);
     Rcpp::traits::input_parameter< int >::type k(kSEXP);
-    rcpp_result_gen = Rcpp::wrap(skip_ngrams(words, n, k));
+    rcpp_result_gen = Rcpp::wrap(skip_ngrams_vectorised(words, stopwords, n, k));
     return rcpp_result_gen;
 END_RCPP
 }