+++ src/skip_ngrams.cpp
@@ -1,12 +1,30 @@
 #include <Rcpp.h>
 using namespace Rcpp;
 
-// Skip n-grams
-// @param n = number of words in an n-gram
-// @param k = max number of words to skip
-// [[Rcpp::export]]
-CharacterVector skip_ngrams(CharacterVector words, int n, int k) {
-  int w = words.size(); // w = number of words
+CharacterVector subset_words(std::deque<std::string>& words, NumericVector indices){
+  CharacterVector output(indices.size());
+
+  for(unsigned int i = 0; i < indices.size(); i++){
+    output[i] = words[indices[i]];
+  }
+
+  return output;
+}
+
+CharacterVector skip_ngrams(CharacterVector words, std::set<std::string>& stopwords, int n, int k) {
+
+  std::deque < std::string > checked_words;
+  std::string holding;
+
+  for(unsigned int i = 0; i < words.size(); i++){
+    if(words[i] != NA_STRING){
+      holding = as<std::string>(words[i]);
+      if(stopwords.find(holding)  == stopwords.end()){
+        checked_words.push_back(holding);
+      }
+    }
+  }
+  int w = checked_words.size(); // w = number of words
   int g = 0; // g = number of n-grams
   for(int i = k; i >= 0; i--) { // i = current iteration of k
     int window = n + n * i - i; // width of n-grams plus skips
@@ -30,8 +48,7 @@ CharacterVector skip_ngrams(CharacterVector words, int n, int k) {
         // Rcpp::Rcout << "j + j * k = " << j + j * k << std::endl;
       }
 
-      CharacterVector words_subset = words[subset];
-
+      CharacterVector words_subset = subset_words(checked_words, subset);
 
       // turn the vector of words into a string
       std::string ngram;
@@ -48,4 +65,29 @@ CharacterVector skip_ngrams(CharacterVector words, int n, int k) {
   }
 
   return ngrams;
+}
+
+//[[Rcpp::export]]
+ListOf<CharacterVector> skip_ngrams_vectorised(ListOf<CharacterVector> words,
+                                               CharacterVector stopwords,
+                                               int n,
+                                               int k){
+
+  // Create output object and set up for further work
+  unsigned int input_size = words.size();
+  List output(input_size);
+
+  // Create stopwords set
+  std::set < std::string > checked_stopwords;
+  for(unsigned int i = 0; i < stopwords.size(); i++){
+    if(stopwords[i] != NA_STRING){
+      checked_stopwords.insert(as<std::string>(stopwords[i]));
+    }
+  }
+
+  for(unsigned int i = 0; i < input_size; i++){
+    output[i] = skip_ngrams(words[i], checked_stopwords, n, k);
+  }
+
+  return output;
 }