+++ NEWS.md
@@ -4,6 +4,7 @@
 - Add the `tokenize_ptb()` function for Penn Tree Bank tokenizations (@jrnold).
 - C++98 has replaced the C++11 code used for ngram generation, widening the range of compilers `tokenizers` supports (#26)
 - If tokenisers fail to generate tokens for a particular entry, they return NA consistently (#33)
+- `tokenize_skip_ngrams` now supports stopwords (#31)
 
 # tokenizers 0.1.4